+++ test/peary/test_peary_protocol/test_peary_protocol_checks.py
@@ -1,0 +1,5 @@
+from peary.peary_protocol import PearyProtocol
+
+
+def test_peary_protocol_class_constants_checks() -> None:
+    assert PearyProtocol.Checks.CHECK_VERSION not in PearyProtocol.Checks.CHECK_NONE