+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python version is now reported with other dependecy versions in the gihub workflow.
 - Added `py.typed` to the source packages to allow for mypy typing across dependencies.
 - Created the supply classes that will be used to access the caribou board supplies.
+- Added an optional checks flag to peary protocol class so intialization checks can
+  be controlled during intialization.
 ### Changed
 - Updated Nox to resuse the virtual environments accross sessions.
 - Reduced the socket timeout for peary protocol from 10s to 1s.